#50 Bulk Upsert into Postgres table
