+++ docker-compose.yml
@@ -40,6 +40,7 @@ services:
     container_name: sync
     environment:
       test_mode: ${test_mode:-0}
+      EXECUTION_ID: ${EXECUTION_ID:-0}
       ORACLE_HOST: ${ORACLE_HOST:-nrcdb03.bcgov}
       ORACLE_PASSWORD: ${ORACLE_PASSWORD:-password}
       ORACLE_PORT: ${ORACLE_PORT:-1543}
